Remove code that is commented out
Ideally we would allow the role user to specify a customer ID for the
CrowdStrike Falcon sensor and simply use a dummy value for testing.
Unfortunately, the customer ID contains a checksum and the falconctl
CLI rejects values with an incorrect checksum; hence I don't know how
to generate a dummy value.